HDFS-5642: cast jReplication to jshort to suppress compiler warning.

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/native/libhdfs/hdfs.c
@@ -908,7 +908,7 @@ hdfsFile hdfsOpenFile(hdfsFS fs, const char* path, int flags,
                     path);
                 goto done;
             }
-            jReplication = jVal.i;
+            jReplication = (jshort)jVal.i;
         }
     }
